+++ StripeCryptoOnramp/StripeCryptoOnramp/Source/Internal/CryptoOnrampAnalyticsEvent.swift
@@ -19,6 +19,7 @@ enum CryptoOnrampOperation: String {
     case submitIdentifiers = "submit_identifiers"
     case presentUserAttestation = "present_user_attestation"
     case presentTermsAndConditionsIfNeeded = "present_terms_and_conditions_if_needed"
+    case presentTermsOfServiceIfNeeded = "present_terms_of_service_if_needed"
     case verifyKycInfo = "verify_kyc_info"
     case verifyIdentity = "verify_identity"
     case registerWalletAddress = "register_wallet_address"
@@ -48,6 +49,8 @@ enum CryptoOnrampAnalyticsEvent {
     case userAttestationCompleted
     case termsAndConditionsStarted
     case termsAndConditionsCompleted
+    case termsOfServiceStarted
+    case termsOfServiceCompleted
     case kycInfoVerificationStarted
     case kycInfoVerificationCompleted
     case walletRegistered(network: String)
@@ -96,6 +99,10 @@ enum CryptoOnrampAnalyticsEvent {
             return "onramp.terms_and_conditions_started"
         case .termsAndConditionsCompleted:
             return "onramp.terms_and_conditions_completed"
+        case .termsOfServiceStarted:
+            return "onramp.terms_of_service_started"
+        case .termsOfServiceCompleted:
+            return "onramp.terms_of_service_completed"
         case .kycInfoVerificationStarted:
             return "onramp.kyc_info_verification_started"
         case .kycInfoVerificationCompleted:
@@ -138,6 +145,8 @@ enum CryptoOnrampAnalyticsEvent {
              .userAttestationCompleted,
              .termsAndConditionsStarted,
              .termsAndConditionsCompleted,
+             .termsOfServiceStarted,
+             .termsOfServiceCompleted,
              .identifierRequirementsRetrieved,
              .userAttestationStarted,
              .kycInfoVerificationStarted,