+++ CHANGELOG.md
@@ -4,7 +4,7 @@ MINOR
 ## X.Y.Z - changes pending release
 ### CryptoOnramp (Alpha)
 * [Added] Added Canada SIN, Colombia NIT, and Philippines TIN values to `IdType`, and added `idType` to the `KycInfo` initializer.
-* [Added] Added `CryptoOnrampCoordinator.presentTermsAndConditionsIfNeeded(from:)` to present and record acceptance of the current terms and conditions when required.
+* [Added] Added `CryptoOnrampCoordinator.presentTermsAndConditionsIfNeeded(from:)` and `presentTermsOfServiceIfNeeded(from:)` to present and record acceptance of current partner terms when required. Both methods return `PartnerTermsResult`.
 
 ## 26.9.0 2026-08-31
 ### Payments